requirements.txt
coverage==4.2
dj-database-url==0.4.0
Django==1.10.2
django-extensions==1.7.4
django-model-utils==2.5
django-simple-history==1.8.1
django-suit==0.2.21
et-xmlfile==1.0.1
gunicorn==19.4.5
jdcal==1.2
model-mommy==1.3.0
openpyxl==2.3.4
Pillow==3.4.2
psycopg2==2.6.1
reportlab==3.3.0
simplejson==3.8.2
six==1.10.0
suds==0.4
whitenoise==2.0.6
